backend/pyproject.toml
[project]
name = "docsnexus"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "datasets>=3.6.0",
    "fastapi>=0.115.12",
    "groq>=0.28.0",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
    "transformers>=4.52.4",
    "uvicorn>=0.34.3",
    "weaviate-client>=4.15.0",
]
